jni_mangle: name + sig args + camelCase convert
This makes a number of changes to the jni_mangle macro...

The macro now accept three arguments:
1. The package name (required / as before)
2. The method name (optional)
3. The method JNI signature (optional)

If two args are given then if the second arg contains brackets it will
be treated as a signature.

The method name argument allows the macro to mangle names that aren't
valid Rust function names (such as `fooBar$Baz`)

The JNI signature argument allows the macro to encode a fully-qualified
name that also allows native method overloads.

Instead of having one special-case for encoding the '$' character, this
also implements general support for encoding any non-ascii characters
in the form `_0xxxx` where `xxxx` is the unicode codepoint in lower-case hex.

This also makes sure to encode `;` and `[` as `_2` and `_3`
respectively (when encoding the arguments of a JNI signature).

The macro allows for the initial function to already have the "system"
ABI be specified - but any other explicit ABI will still trigger an
error.

If no method name is explicitly specified then instead of directly
taking the Rust function name as the name, the macro will first
convert a lower_snake_case name to camelCase, preserving any leading or
trailing underscores.

Note: no camelCase transform is done if the Rust function has any
upper-case characters, so a name like `fooBar_Baz` would be used as-is,
and for anything extra-weird the name can be given to the macro as a
string literal.

When validating the package name the macro will now throw an error if
the package name starts with a '.' or contains any double dots.