jni-macros: bind_java_type, jni_sig, jni_str + native_method macros
# `bind_java_type` macro

This is a comprehensive re-write of the `bind_class` macro (that I first
attempted to write as a declarative, macro_rules macro).

Shorthand usage of the macro looks like:

```
bind_java_type! { RustType => test.JavaType }
```

Typical usage of the macro looks like:

```
    bind_java_type! {
        rust_type = RustType,
        java_type = test.JavaType,
        type_map {
            OtherRustType => test.OtherJavaType
        },
        constructors {
            new(),
            with_value(value: jint),
        },
        methods = {
            get_name() -> JString,
            set_value(val: jint) -> void,
            do_something(arg: OtherRustType) -> jint[],
        },
        fields = {
            count: jint
        }
    }
```

# `jni_sig` macro

This macro enables compile-time parsing of a method or field signature
and outputs a `MethodSignature` or `FieldSignature` struct that includes
a literal signature string (encoded as MUTF-8) and maps method
arguments, return values and field types into `JavaType` enum variants.

The macro can either be passed a standard JNI signature as input (e.g.
"(ILjava/lang/String;)Z") or it accepts the same signature syntax as the
`bind_java_type` macro, such as `(jint, JString) -> jboolean`

# `jni_sig_str`, `jni_sig_cstr` + `jni_sig_jstr` macros

Accept the same input as `jni_sig` but instead of outputting a struct
they just output a string literal for the signature, either as a plain
`&'static str`, a `&'static CStr` or a `&'static JNIStr` respectively.

# `jni_str` + `jni_cstr` macros

Act like the `std` `concat!()` macro, except they concatenate their
arguments into a MUTF-8 encoded `&JNIStr` or `&CStr` literal
respectively.

# native_method

This makes it possible to construct an `env::NativeMethod` with a
guarantee that the function pointer matches the JNI signature.

This has a syntax consistent with `bind_java_type!` and `jni_sig!`
with a long-form `prop = <val>,` config syntax and a short-form
signature syntax.

For example:

```
let method = native_method! {
    static extern fn MyRustType::my_native_method(arg: jint) -> bool
};
```

In this case `my_native_method` is mapped to "myNativeMethod" as the
Java method name. The signature would be "(I)Z" and the implementation
function would be `MyRustType::my_native_method`. If it was an instance
method then `MyRustType` would also be used for the `this` argument.